+++ src/GraphQl/Serializer/ItemNormalizer.php
@@ -73,7 +73,7 @@ final class ItemNormalizer extends BaseItemNormalizer
 
         $data = parent::normalize($object, $format, $context);
         if (!\is_array($data)) {
-            throw new UnexpectedValueException('Expected data to be an array');
+            throw new UnexpectedValueException('Expected data to be an array.');
         }
 
         $data[self::ITEM_RESOURCE_CLASS_KEY] = $this->getObjectClass($object);