+++ tests/GraphQl/Action/EntrypointActionTest.php
@@ -17,6 +17,8 @@ use ApiPlatform\Core\GraphQl\Action\EntrypointAction;
 use ApiPlatform\Core\GraphQl\Action\GraphiQlAction;
 use ApiPlatform\Core\GraphQl\Action\GraphQlPlaygroundAction;
 use ApiPlatform\Core\GraphQl\ExecutorInterface;
+use ApiPlatform\Core\GraphQl\Serializer\Exception\ErrorNormalizer;
+use ApiPlatform\Core\GraphQl\Serializer\Exception\HttpExceptionNormalizer;
 use ApiPlatform\Core\GraphQl\Type\SchemaBuilderInterface;
 use GraphQL\Executor\ExecutionResult;
 use GraphQL\Type\Schema;
@@ -27,6 +29,7 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\RouterInterface;
+use Symfony\Component\Serializer\Serializer;
 use Twig\Environment as TwigEnvironment;
 
 /**
@@ -37,7 +40,7 @@ class EntrypointActionTest extends TestCase
     /**
      * Hack to avoid transient failing test because of Date header.
      */
-    private function assertEqualsWithoutDateHeader(JsonResponse $expected, Response $actual)
+    private function assertEqualsWithoutDateHeader(JsonResponse $expected, Response $actual): void
     {
         $expected->headers->remove('Date');
         $actual->headers->remove('Date');
@@ -53,7 +56,7 @@ class EntrypointActionTest extends TestCase
         $this->assertInstanceOf(Response::class, $mockedEntrypoint($request));
     }
 
-    public function testGetAction()
+    public function testGetAction(): void
     {
         $request = new Request(['query' => 'graphqlQuery', 'variables' => '["graphqlVariable"]', 'operation' => 'graphqlOperationName']);
         $request->setRequestFormat('json');
@@ -62,7 +65,7 @@ class EntrypointActionTest extends TestCase
         $this->assertEqualsWithoutDateHeader(new JsonResponse(['GraphQL']), $mockedEntrypoint($request));
     }
 
-    public function testPostRawAction()
+    public function testPostRawAction(): void
     {
         $request = new Request(['variables' => '["graphqlVariable"]', 'operation' => 'graphqlOperationName'], [], [], [], [], [], 'graphqlQuery');
         $request->setFormat('graphql', 'application/graphql');
@@ -73,7 +76,7 @@ class EntrypointActionTest extends TestCase
         $this->assertEqualsWithoutDateHeader(new JsonResponse(['GraphQL']), $mockedEntrypoint($request));
     }
 
-    public function testPostJsonAction()
+    public function testPostJsonAction(): void
     {
         $request = new Request([], [], [], [], [], [], '{"query": "graphqlQuery", "variables": "[\"graphqlVariable\"]", "operation": "graphqlOperationName"}');
         $request->setMethod('POST');
@@ -86,7 +89,7 @@ class EntrypointActionTest extends TestCase
     /**
      * @dataProvider multipartRequestProvider
      */
-    public function testMultipartRequestAction(?string $operations, ?string $map, array $files, array $variables, Response $expectedResponse)
+    public function testMultipartRequestAction(?string $operations, ?string $map, array $files, array $variables, Response $expectedResponse): void
     {
         $requestParams = [];
         if ($operations) {
@@ -144,82 +147,82 @@ class EntrypointActionTest extends TestCase
                 '{"file": ["variables.file"]}',
                 ['file' => $file],
                 ['file' => $file],
-                new Response('{"errors":[{"message":"GraphQL multipart request does not respect the specification.","extensions":{"category":"user"}}]}'),
+                new Response('{"errors":[{"message":"GraphQL multipart request does not respect the specification.","extensions":{"category":"user","status":400}}]}'),
             ],
             'upload without providing map' => [
                 '{"query": "graphqlQuery", "variables": {"file": null}, "operation": "graphqlOperationName"}',
                 null,
                 ['file' => $file],
                 ['file' => null],
-                new Response('{"errors":[{"message":"GraphQL multipart request does not respect the specification.","extensions":{"category":"user"}}]}'),
+                new Response('{"errors":[{"message":"GraphQL multipart request does not respect the specification.","extensions":{"category":"user","status":400}}]}'),
             ],
             'upload with invalid json' => [
                 '{invalid}',
                 '{"file": ["file"]}',
                 ['file' => $file],
                 ['file' => null],
-                new Response('{"errors":[{"message":"GraphQL data is not valid JSON.","extensions":{"category":"user"}}]}'),
+                new Response('{"errors":[{"message":"GraphQL data is not valid JSON.","extensions":{"category":"user","status":400}}]}'),
             ],
             'upload with invalid map JSON' => [
                 '{"query": "graphqlQuery", "variables": {"file": null}, "operation": "graphqlOperationName"}',
                 '{invalid}',
                 ['file' => $file],
                 ['file' => null],
-                new Response('{"errors":[{"message":"GraphQL multipart request map is not valid JSON.","extensions":{"category":"user"}}]}'),
+                new Response('{"errors":[{"message":"GraphQL multipart request map is not valid JSON.","extensions":{"category":"user","status":400}}]}'),
             ],
             'upload with no file' => [
                 '{"query": "graphqlQuery", "variables": {"file": null}, "operation": "graphqlOperationName"}',
                 '{"file": ["file"]}',
                 [],
                 ['file' => null],
-                new Response('{"errors":[{"message":"GraphQL multipart request file has not been sent correctly.","extensions":{"category":"user"}}]}'),
+                new Response('{"errors":[{"message":"GraphQL multipart request file has not been sent correctly.","extensions":{"category":"user","status":400}}]}'),
             ],
             'upload with wrong map' => [
                 '{"query": "graphqlQuery", "variables": {"file": null}, "operation": "graphqlOperationName"}',
                 '{"file": ["file"]}',
                 ['file' => $file],
                 ['file' => null],
-                new Response('{"errors":[{"message":"GraphQL multipart request path in map is invalid.","extensions":{"category":"user"}}]}'),
+                new Response('{"errors":[{"message":"GraphQL multipart request path in map is invalid.","extensions":{"category":"user","status":400}}]}'),
             ],
             'upload when variable path does not exist' => [
                 '{"query": "graphqlQuery", "variables": {"file": null}, "operation": "graphqlOperationName"}',
                 '{"file": ["variables.wrong"]}',
                 ['file' => $file],
                 ['file' => null],
-                new Response('{"errors":[{"message":"GraphQL multipart request path in map does not match the variables.","extensions":{"category":"user"}}]}'),
+                new Response('{"errors":[{"message":"GraphQL multipart request path in map does not match the variables.","extensions":{"category":"user","status":400}}]}'),
             ],
         ];
     }
 
-    public function testBadContentTypePostAction()
+    public function testBadContentTypePostAction(): void
     {
         $request = new Request();
         $request->setMethod('POST');
         $request->headers->set('Content-Type', 'application/xml');
         $mockedEntrypoint = $this->getEntrypointAction();
 
-        $this->assertEquals(400, $mockedEntrypoint($request)->getStatusCode());
-        $this->assertEquals('{"errors":[{"message":"GraphQL query is not valid.","extensions":{"category":"user"}}]}', $mockedEntrypoint($request)->getContent());
+        $this->assertEquals(200, $mockedEntrypoint($request)->getStatusCode());
+        $this->assertEquals('{"errors":[{"message":"GraphQL query is not valid.","extensions":{"category":"user","status":400}}]}', $mockedEntrypoint($request)->getContent());
     }
 
-    public function testBadMethodAction()
+    public function testBadMethodAction(): void
     {
         $request = new Request();
         $request->setMethod('PUT');
         $mockedEntrypoint = $this->getEntrypointAction();
 
-        $this->assertEquals(400, $mockedEntrypoint($request)->getStatusCode());
-        $this->assertEquals('{"errors":[{"message":"GraphQL query is not valid.","extensions":{"category":"user"}}]}', $mockedEntrypoint($request)->getContent());
+        $this->assertEquals(200, $mockedEntrypoint($request)->getStatusCode());
+        $this->assertEquals('{"errors":[{"message":"GraphQL query is not valid.","extensions":{"category":"user","status":400}}]}', $mockedEntrypoint($request)->getContent());
     }
 
-    public function testBadVariablesAction()
+    public function testBadVariablesAction(): void
     {
         $request = new Request(['query' => 'graphqlQuery', 'variables' => 'graphqlVariable', 'operation' => 'graphqlOperationName']);
         $request->setRequestFormat('json');
         $mockedEntrypoint = $this->getEntrypointAction();
 
-        $this->assertEquals(400, $mockedEntrypoint($request)->getStatusCode());
-        $this->assertEquals('{"errors":[{"message":"GraphQL variables are not valid JSON.","extensions":{"category":"user"}}]}', $mockedEntrypoint($request)->getContent());
+        $this->assertEquals(200, $mockedEntrypoint($request)->getStatusCode());
+        $this->assertEquals('{"errors":[{"message":"GraphQL variables are not valid JSON.","extensions":{"category":"user","status":400}}]}', $mockedEntrypoint($request)->getContent());
     }
 
     private function getEntrypointAction(array $variables = ['graphqlVariable']): EntrypointAction
@@ -228,8 +231,14 @@ class EntrypointActionTest extends TestCase
         $schemaBuilderProphecy = $this->prophesize(SchemaBuilderInterface::class);
         $schemaBuilderProphecy->getSchema()->willReturn($schema->reveal());
 
+        $normalizer = new Serializer([
+            new HttpExceptionNormalizer(),
+            new ErrorNormalizer(),
+        ]);
+
         $executionResultProphecy = $this->prophesize(ExecutionResult::class);
         $executionResultProphecy->toArray(false)->willReturn(['GraphQL']);
+        $executionResultProphecy->setErrorFormatter([$normalizer, 'normalize'])->willReturn($executionResultProphecy);
         $executorProphecy = $this->prophesize(ExecutorInterface::class);
         $executorProphecy->executeQuery(Argument::is($schema->reveal()), 'graphqlQuery', null, null, $variables, 'graphqlOperationName')->willReturn($executionResultProphecy->reveal());
 
@@ -239,6 +248,6 @@ class EntrypointActionTest extends TestCase
         $graphiQlAction = new GraphiQlAction($twigProphecy->reveal(), $routerProphecy->reveal(), true);
         $graphQlPlaygroundAction = new GraphQlPlaygroundAction($twigProphecy->reveal(), $routerProphecy->reveal(), true);
 
-        return new EntrypointAction($schemaBuilderProphecy->reveal(), $executorProphecy->reveal(), $graphiQlAction, $graphQlPlaygroundAction, false, true, true, 'graphiql');
+        return new EntrypointAction($schemaBuilderProphecy->reveal(), $executorProphecy->reveal(), $graphiQlAction, $graphQlPlaygroundAction, $normalizer, false, true, true, 'graphiql');
     }
 }